Move data fetching + transformation for sequences over time to new file
